+++ ft_strcmp.c
@@ -6,13 +6,15 @@
 /*   By: labdello <labdello@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/06 15:23:03 by labdello          #+#    #+#             */
-/*   Updated: 2024/04/06 15:52:49 by labdello         ###   ########.fr       */
+/*   Updated: 2024/04/07 13:03:10 by labdello         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int	ft_strcmp(char *s1, char *s2)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] != '\0' || s2[i] != '\0')
