+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: labdello <labdello@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/06 13:28:39 by labdello          #+#    #+#             */
-/*   Updated: 2024/04/06 14:33:41 by labdello         ###   ########.fr       */
+/*   Updated: 2024/04/07 13:04:24 by labdello         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,6 @@
 # include <string.h>
 # include <stdlib.h>
 
-int	ft_strlen(char *str);
+size_t	ft_strlen(char *str);
 
 #endif
